Reorganise home and landing pages
