+++ src/app/components/navbar/navbar.component.scss
@@ -25,6 +25,12 @@
   text-decoration: none;
 }
 
+.brand-text {
+  display: flex;
+  flex-direction: column;
+  gap: 0.1rem;
+}
+
 .brand-mark {
   width: 2.5rem;
   height: 2.5rem;
@@ -89,6 +95,12 @@
 .nav-links {
   display: flex;
   align-items: center;
+  gap: 1rem;
+}
+
+.nav-link-group {
+  display: flex;
+  align-items: center;
   gap: 0.5rem;
 }
 
@@ -117,6 +129,46 @@
   box-shadow: inset 0 0 0 1px rgba(20, 184, 166, 0.35);
 }
 
+.sign-out {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.6rem 1.15rem;
+  border-radius: 0.75rem;
+  border: none;
+  background: linear-gradient(135deg, #0f766e, #0ea5a4);
+  color: #fff;
+  font-weight: 600;
+  box-shadow: 0 12px 25px -18px rgba(14, 165, 233, 0.6);
+  cursor: pointer;
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+  margin-left: clamp(0.75rem, 3vw, 2rem);
+}
+
+.sign-out:hover,
+.sign-out:focus-visible {
+  transform: translateY(-1px);
+  box-shadow: 0 14px 30px -16px rgba(15, 118, 110, 0.55);
+  outline: none;
+}
+
+.sign-out:disabled {
+  cursor: progress;
+  opacity: 0.75;
+  transform: none;
+  box-shadow: none;
+}
+
+.sign-out-error {
+  margin: 0.5rem auto 0;
+  max-width: 1200px;
+  padding: 0.75rem 1rem;
+  border-radius: 0.75rem;
+  background: rgba(248, 113, 113, 0.12);
+  color: rgba(185, 28, 28, 0.95);
+  font-size: 0.85rem;
+}
+
 @media (max-width: 900px) {
   .nav-inner {
     flex-wrap: wrap;
@@ -139,9 +191,22 @@
     display: flex;
   }
 
+  .nav-link-group {
+    flex-direction: column;
+    width: 100%;
+    gap: 0.35rem;
+  }
+
   .nav-link {
     width: 100%;
     justify-content: flex-start;
+  }
+
+  .sign-out {
+    width: 100%;
+    justify-content: center;
+    margin-top: 0.75rem;
+    margin-left: 0;
   }
 }
