+++ src/app/pages/main-page/main-page.component.scss
